Moved shared functionality from FileQueue and Queue into WithFiles mixin.
Refactored flushFilesWhenSettled observer into a plain flush function. Call flush when necessary.
