+++ addon/queue.js
@@ -1,13 +1,12 @@
 import { A } from '@ember/array';
 import EmberObject, {
-  observer,
-  computed,
   set,
   get
 } from '@ember/object';
 import { next } from '@ember/runloop';
 import File from './file';
-import sumBy from './computed/sum-by';
+import WithFiles from './mixins/with-files';
+
 
 /**
   The Queue is a collection of files that
@@ -20,7 +19,7 @@ import sumBy from './computed/sum-by';
   @class Queue
   @extends Ember.Object
  */
-export default EmberObject.extend({
+export default EmberObject.extend(WithFiles, {
 
   init() {
     set(this, 'files', A());
@@ -130,78 +129,5 @@ export default EmberObject.extend({
     @type File[]
     @default []
    */
-  files: null,
-
-  /**
-    Flushes the `files` property when they have settled. This
-    will only flush files when all files have arrived at a terminus
-    of their state chart.
-
-    ```
-        .------.     .---------.     .--------.
-    o--| queued |-->| uploading |-->| uploaded |
-        `------`     `---------`     `--------`
-           ^              |    .-------.
-           |              |`->| aborted |
-           |              |    `-------`
-           |  .------.    |    .---------.
-           `-| failed |<-` `->| timed_out |-.
-           |  `------`         `---------`  |
-           `-------------------------------`
-    ```
-
-    Files *may* be requeued by the user in the `failed` or `timed_out`
-    states.
-
-    @private
-   */
-  flushFilesWhenSettled: observer('files.@each.state', function () { // eslint-disable-line ember/no-observers
-    let files = get(this, 'files');
-    let allFilesHaveSettled = files.every(function (file) {
-      return ['uploaded', 'aborted'].indexOf(file.state) !== -1;
-    });
-
-    if (files.length === 0) { return; }
-
-    if (allFilesHaveSettled) {
-      get(this, 'files').forEach((file) => set(file, 'queue', null));
-      set(this, 'files', A());
-    }
-  }),
-
-  /**
-    The aggregate size (in bytes) of all files in the queue.
-
-    @accessor size
-    @readonly
-    @type number
-    @default 0
-   */
-  size: sumBy('files', 'size'),
-
-  /**
-    The aggregate amount of bytes that have been uploaded
-    to the server for all files in the queue.
-
-    @accessor loaded
-    @readonly
-    @type number
-    @default 0
-   */
-  loaded: sumBy('files', 'loaded'),
-
-  /**
-    The current upload progress of the queue, as a number from 0 to 100.
-
-    @accessor progress
-    @readonly
-    @type number
-    @default 0
-   */
-  progress: computed('size', 'loaded', {
-    get() {
-      let percent = (get(this, 'loaded') / get(this, 'size')) || 0;
-      return Math.floor(percent * 100);
-    }
-  })
+  files: null
 });
